refactored tools.jsx
